Alternate trianglulation in Studio [WIP]
